requirements.txt
anyio==3.3.1
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
autonormalize==1.0.1
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
boto3==1.18.26
botocore==1.21.26
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
dask==2021.8.1
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
distributed==2021.8.1
entrypoints==0.3
featuretools==0.27.0
filelock==3.0.12
Flask==2.0.1
fsspec==2021.8.1
haversine==2.4.0
HeapDict==1.0.1
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.13
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
kiwisolver==1.3.2
llvmlite==0.37.0
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.4
numba==0.54.1
numpy==1.20.3
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
phone-iso3166==0.3.7
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pynndescent==0.5.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzipcode==3.0.1
pyzmq==22.3.0
regex==2021.8.28
requests==2.26.0
requests-unixsocket==0.2.0
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
structlog==21.1.0
tblib==1.7.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toolz==0.11.1
torch==1.9.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
umap-learn==0.5.1
urllib3==1.26.6
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
xgboost==1.4.2
zict==2.0.0
Flask==2.0.1
